Extract supplement impact module
